823: Refactor word_admin
