+++ lunes_cms/cmsv2/admins/word_admin.py
@@ -1,119 +1,26 @@
 from __future__ import absolute_import, annotations, unicode_literals
 
 from datetime import date
-from typing import TYPE_CHECKING
 
-from django.contrib import admin
-from django.urls import reverse
-from django.utils.html import escape, format_html
 from django.utils.safestring import mark_safe, SafeString
 from django.utils.translation import gettext_lazy as _
 
 from lunes_cms.cmsv2.admins.base import BaseAdmin
+from lunes_cms.cmsv2.admins.word_admin_asset_widgets import WordAdminAssetWidgetsMixin
+from lunes_cms.cmsv2.admins.word_admin_list_renderers import (
+    WordAdminListRenderersMixin,
+)
 from lunes_cms.cmsv2.admins.word_filters import (
     HasCompleteExampleSentenceFilter,
     HasImageFilter,
     MigratedFilter,
     UnitOrJobDropdownFilter,
 )
-from lunes_cms.cmsv2.models import AlternativeWord, Word
-from lunes_cms.cmsv2.models.static import CheckStatus
-from lunes_cms.cmsv2.models.unit import UnitWordRelation
-from lunes_cms.cmsv2.utils import (
-    cache_busted_url,
-    example_sentence_generate_html,
-    get_image_tag,
-    is_not_blank,
-)
-from lunes_cms.core import settings
-
-if TYPE_CHECKING:
-    # `_StrOrPromise` only exists in django-stubs, not at runtime.
-    from django.utils.functional import _StrOrPromise
+from lunes_cms.cmsv2.admins.word_inlines import AlternativeWordInline, UnitInline
+from lunes_cms.cmsv2.models import Word
 
 
-class AlternativeWordInline(admin.TabularInline):
-    """
-    Inline admin for the AlternativeWord model.
-
-    This inline allows editing alternative words directly from the Word admin page.
-    """
-
-    model = AlternativeWord
-    extra = 1
-    can_delete = False
-    verbose_name = _("alternative word")
-    verbose_name_plural = _("So heißt das auch")
-    fields = [
-        "grammatical_gender",
-        "singular_article",
-        "alt_word",
-        "plural_article",
-        "plural",
-        "action_buttons",
-    ]
-    readonly_fields = ["action_buttons"]
-
-    def action_buttons(self, obj: AlternativeWord) -> SafeString:
-        """
-        Render buttons which instantly add, save or delete the alternative
-        word of the row, so no separate save of the whole word is needed.
-
-        Args:
-            obj: The alternative word object
-
-        Returns:
-            str: HTML markup for the action buttons
-        """
-        if not obj.pk:
-            return format_html(
-                '<button type="button" class="add-alternative-word-btn" title="{}">'
-                '<span class="alternative-word-add">+</span></button>',
-                _("Add alternative word"),
-            )
-        return format_html(
-            '<button type="button" class="save-alternative-word-btn" '
-            'data-alternative-word-id="{}" title="{}">'
-            '<span class="alternative-word-save">✓</span></button>'
-            '<button type="button" class="delete-alternative-word-btn" '
-            'data-alternative-word-id="{}" title="{}">'
-            '<span class="alternative-word-delete">×</span></button>',
-            obj.pk,
-            _("Save alternative word"),
-            obj.pk,
-            _("Delete alternative word"),
-        )
-
-    action_buttons.short_description = ""  # type: ignore[attr-defined]
-
-
-class UnitInline(admin.TabularInline):
-    """
-    Inline admin for UnitWordRelation model.
-
-    This inline allows editing unit-word relationships directly from the Word admin page,
-    including the ability to add/edit images and audio for each unit-word relation.
-    """
-
-    model = UnitWordRelation
-    extra = 1
-    autocomplete_fields = ["unit"]
-    fields = [
-        "unit",
-        "image_with_controls",
-        "example_sentence",
-        "example_sentence_generate",
-        "example_sentence_check_status",
-        "example_sentence_audio_player",
-    ]
-    readonly_fields = [
-        "image_with_controls",
-        "example_sentence_generate",
-        "example_sentence_audio_player",
-    ]
-
-
-class WordAdmin(BaseAdmin):
+class WordAdmin(BaseAdmin, WordAdminAssetWidgetsMixin, WordAdminListRenderersMixin):
     """
     Admin interface for the Word model.
 
@@ -242,304 +149,6 @@ class WordAdmin(BaseAdmin):
         ]
         css = {"all": ["css/asset_manager.css", "css/audio_player.css"]}
 
-    def _render_regenerate_widget(  # pylint: disable=too-many-arguments
-        self,
-        *,
-        asset_type: str,
-        generate_url: str,
-        store_url: str,
-        text_field: str,
-        text_value: _StrOrPromise,
-        store_field: str,
-        current_preview: _StrOrPromise,
-        generate_label: _StrOrPromise,
-        regenerate_label: _StrOrPromise,
-        with_additional_info: bool = False,
-        spoken_text: str = "",
-    ) -> SafeString:
-        """
-        Render an inline (re)generation widget for the change page.
-
-        The widget generates a new audio file or image via OpenAI without
-        leaving the page, shows the current and the newly generated version
-        side by side for comparison, and lets the user keep the new one or
-        discard it and keep the current one.
-
-        ``spoken_text`` is shown above the buttons as a reminder to save first.
-        """
-        spoken_text_html: _StrOrPromise = ""
-        if spoken_text:
-            spoken_text_html = format_html(
-                '<div class="regen-spoken-text">{} „{}“</div>',
-                _("Will be spoken:"),
-                spoken_text,
-            )
-
-        additional_info_html: _StrOrPromise = ""
-        if with_additional_info:
-            additional_info_html = format_html(
-                '<div class="regen-allow-text-row">'
-                '<label><input type="checkbox" class="regen-allow-text"> {}</label>'
-                "</div>"
-                '<div class="regen-additional-info-row">'
-                "<label>{} "
-                '<input type="text" class="regen-additional-info"></label>'
-                "</div>",
-                _(
-                    "Allow text/numbers in the image (e.g. for a receipt or a clock face)"
-                ),
-                _("Additional info (optional)"),
-            )
-
-        return format_html(
-            '<div class="inline-regenerate" data-asset-type="{asset_type}" '
-            'data-generate-url="{generate_url}" data-store-url="{store_url}" '
-            'data-text-field="{text_field}" data-text="{text_value}" '
-            'data-store-field="{store_field}">'
-            '<div class="regen-compare">'
-            '<div class="regen-col regen-current">'
-            '<div class="regen-col-label">{current_label}</div>'
-            '<div class="regen-current-preview">{current_preview}</div>'
-            "</div>"
-            '<div class="regen-col regen-new">'
-            '<div class="regen-col-label">{new_label}</div>'
-            '<div class="regen-new-empty">{empty_label}</div>'
-            '<div class="regen-new-preview"></div>'
-            "</div>"
-            "</div>"
-            "{spoken_text_html}"
-            "{additional_info_html}"
-            '<div class="regen-toolbar">'
-            '<button type="button" class="btn btn-primary btn-sm regen-generate-btn" '
-            'data-regenerate-label="{regenerate_label}">{generate_label}</button>'
-            '<span class="regen-spinner spinner-border spinner-border-sm is-hidden"></span>'
-            '<span class="regen-message"></span>'
-            "</div>"
-            '<div class="regen-decision is-hidden">'
-            '<button type="button" class="btn btn-success btn-sm regen-keep-btn">{keep_label}</button> '
-            '<button type="button" class="btn btn-secondary btn-sm regen-discard-btn">{discard_label}</button>'
-            "</div>"
-            "</div>",
-            asset_type=asset_type,
-            generate_url=generate_url,
-            store_url=store_url,
-            text_field=text_field,
-            text_value=text_value,
-            store_field=store_field,
-            current_label=_("Current"),
-            new_label=_("New"),
-            empty_label=_("Not generated yet"),
-            current_preview=current_preview,
-            spoken_text_html=spoken_text_html,
-            additional_info_html=additional_info_html,
-            generate_label=generate_label,
-            regenerate_label=regenerate_label,
-            keep_label=_("Keep new"),
-            discard_label=_("Discard new"),
-        )
-
-    def audio_generate(self, obj: Word) -> _StrOrPromise:
-        """
-        Generate HTML for the inline audio (re)generation widget.
-
-        Args:
-            obj: The word object
-
-        Returns:
-            str: HTML markup for the audio generation widget
-        """
-        if not obj.pk:
-            return _("Save to enable audio generation.")
-        current_preview: _StrOrPromise
-        if obj.audio:
-            current_preview = format_html(
-                "<audio controls src='{}'></audio>", cache_busted_url(obj.audio)
-            )
-        else:
-            current_preview = _("No audio yet.")
-        return self._render_regenerate_widget(
-            asset_type="audio",
-            generate_url=reverse("cmsv2:word_generate_audio_via_openai", args=[obj.pk]),
-            store_url=reverse(
-                "cmsv2:word_store_generated_audio_permanently", args=[obj.pk]
-            ),
-            text_field="word_text",
-            text_value=obj.text_for_audio_generation(),
-            store_field="temp_audio_filename",
-            current_preview=current_preview,
-            generate_label=_("Generate audio"),
-            regenerate_label=_("Regenerate audio"),
-            spoken_text=obj.text_for_audio_generation(),
-        )
-
-    audio_generate.short_description = _("Audio Generation")  # type: ignore[attr-defined]
-
-    def audio_player(self, obj: Word) -> _StrOrPromise:
-        """
-        Generate HTML for the audio player preview.
-
-        Args:
-            obj: The word object
-
-        Returns:
-            str: HTML markup for the audio player
-        """
-        if obj.audio:
-            return format_html(
-                "<audio controls id='audio_preview_player' src='{}'></audio>",
-                cache_busted_url(obj.audio),
-            )
-        return "No audio file uploaded."
-
-    audio_player.short_description = "Audio Preview"  # type: ignore[attr-defined]
-
-    def example_sentence_generate(self, obj: Word) -> _StrOrPromise:
-        """
-        Generate HTML for the inline example sentence generation widget.
-
-        Args:
-            obj: The word object
-
-        Returns:
-            str: HTML markup for the example sentence generation widget
-        """
-        if obj.pk:
-            return example_sentence_generate_html(
-                generate_url=reverse(
-                    "cmsv2:word_generate_example_sentence_via_openai", args=[obj.pk]
-                ),
-                store_url=reverse(
-                    "cmsv2:word_store_generated_example_sentence", args=[obj.pk]
-                ),
-                target="id_example_sentence",
-            )
-        return _("Save to enable example sentence generation.")
-
-    example_sentence_generate.short_description = _("Example Sentence Generation")  # type: ignore[attr-defined]
-
-    def example_sentence_audio_generate(self, obj: Word) -> _StrOrPromise:
-        """
-        Generate HTML for the inline example sentence audio (re)generation widget.
-
-        Args:
-            obj: The word object
-
-        Returns:
-            str: HTML markup for the audio generation widget
-        """
-        if not obj.pk or not is_not_blank(obj.example_sentence):
-            return _("Save to enable audio generation.")
-        current_preview: _StrOrPromise
-        if obj.example_sentence_audio:
-            current_preview = format_html(
-                "<audio controls src='{}'></audio>",
-                cache_busted_url(obj.example_sentence_audio),
-            )
-        else:
-            current_preview = _("No audio yet.")
-        return self._render_regenerate_widget(
-            asset_type="audio",
-            generate_url=reverse(
-                "cmsv2:word_generate_example_sentence_audio_via_openai", args=[obj.pk]
-            ),
-            store_url=reverse(
-                "cmsv2:word_store_generated_example_sentence_audio_permanently",
-                args=[obj.pk],
-            ),
-            text_field="example_sentence_text",
-            text_value=obj.example_sentence,
-            store_field="temp_audio_filename",
-            current_preview=current_preview,
-            generate_label=_("Generate audio"),
-            regenerate_label=_("Regenerate audio"),
-        )
-
-    example_sentence_audio_generate.short_description = _("Example Sentence Audio Generation")  # type: ignore[attr-defined]
-
-    def example_sentence_audio_player(self, obj: Word) -> _StrOrPromise:
-        """
-        Generate HTML for the example sentence audio player preview.
-
-        Args:
-            obj: The word object
-
-        Returns:
-            str: HTML markup for the audio player
-        """
-        if obj.example_sentence_audio:
-            return format_html(
-                "<audio controls id='example_sentence_audio_preview_player' src='{}'></audio>",
-                cache_busted_url(obj.example_sentence_audio),
-            )
-        return "No audio file uploaded."
-
-    example_sentence_audio_player.short_description = "Example Sentence Audio Preview"  # type: ignore[attr-defined]
-
-    def image_tag(self, obj: Word) -> _StrOrPromise:
-        """
-        Generate HTML for displaying the word's image with hover-to-enlarge functionality.
-
-        Args:
-            obj: The word object
-
-        Returns:
-            str: HTML markup for the image with hover overlay
-        """
-        if obj.image:
-            return format_html(
-                """<div class="image-hover-container">
-                    <a href="{}" target="_blank">{}</a>
-                    <div class="image-hover-overlay">
-                        <img src="{}" alt="{}">
-                    </div>
-                </div>""",
-                f"{settings.MEDIA_URL}{obj.image}",
-                mark_safe(get_image_tag(obj.image, width=120)),
-                f"{settings.MEDIA_URL}{obj.image}",
-                escape(obj.word),
-            )
-        return "No image uploaded."
-
-    image_tag.short_description = _("Image Preview")  # type: ignore[attr-defined]
-
-    def image_generate(self, obj: Word) -> _StrOrPromise:
-        """
-        Generate HTML for the inline image (re)generation widget.
-
-        Args:
-            obj: The word object
-
-        Returns:
-            str: HTML markup for the image generation widget
-        """
-        if not obj.pk:
-            return _("Save to enable image generation.")
-        current_preview: _StrOrPromise
-        if obj.image:
-            current_preview = format_html(
-                '<img src="{}" alt="{}" style="max-width: min(200px, 100%);">',
-                f"{settings.MEDIA_URL}{obj.image}",
-                obj.word,
-            )
-        else:
-            current_preview = _("No image yet.")
-        return self._render_regenerate_widget(
-            asset_type="image",
-            generate_url=reverse("cmsv2:generate_image_via_openai"),
-            store_url=reverse(
-                "cmsv2:word_store_generated_image_permanently", args=[obj.pk]
-            ),
-            text_field="word_text",
-            text_value=obj.word,
-            store_field="temp_filename",
-            current_preview=current_preview,
-            generate_label=_("Generate image"),
-            regenerate_label=_("Regenerate image"),
-            with_additional_info=True,
-        )
-
-    image_generate.short_description = _("Image Generation")  # type: ignore[attr-defined]
-
     def creator_group(self, obj: Word) -> str | None:
         """
         Determine the creator group for display in the admin interface.
@@ -558,376 +167,6 @@ class WordAdmin(BaseAdmin):
         return None
 
     creator_group.short_description = _("group")  # type: ignore[attr-defined]
-
-    def list_audio(self, obj: Word) -> SafeString:
-        """
-        Generate HTML for displaying the word's audio with controls in the admin list view.
-
-        This method creates HTML that includes an audio player and buttons for adding,
-        replacing, or deleting the audio file, as well as a dropdown for the audio check status.
-
-        Args:
-            obj: The word object
-
-        Returns:
-            str: HTML markup for displaying the word's audio with controls
-        """
-        audio_html = ""
-        if obj.audio:
-            audio_html = f"""
-            <div class="audio-player-container">
-                <audio class="minimal-audio-player"><source src="{cache_busted_url(obj.audio)}" type="audio/mpeg"></audio>
-                <div class="play-btn">
-                    <div>
-                        <i class="fas fa-play"></i>
-                    </div>
-                </div>
-                <div class="pause-btn" style="display: none;">
-                    <div>
-                        <i class="fas fa-pause"></i>
-                    </div>
-                </div>
-            </div>
-            """
-
-        controls_html = f"""
-        <div class="audio-asset-controls" data-word-id="{obj.id}">
-            <button type="button" class="add-audio-btn" style="display: {'none' if obj.audio else 'inline-flex'};">
-                <span class="audio-add">+</span>
-            </button>
-            <button type="button" class="replace-audio-btn" style="display: {'inline-flex' if obj.audio else 'none'};">
-                <span class="audio-replace">↻</span>
-            </button>
-            <button type="button" class="delete-audio-btn" style="display: {'inline-flex' if obj.audio else 'none'};">
-                <span class="audio-delete">×</span>
-            </button>
-            <input type="file" class="audio-file-input" style="display: none;" accept="audio/*">
-        </div>
-        """
-
-        word_audio_container = (
-            f'<div class="word-audio-container">{audio_html}{controls_html}</div>'
-        )
-
-        options = ""
-        for value, display in CheckStatus.choices:
-            selected = "selected" if obj.audio_check_status == value else ""
-            options += f'<option value="{value}" {selected}>{display}</option>'
-
-        html = word_audio_container
-        if obj.audio:
-            html += f"""
-            <select name="audio_check_status_{obj.id}" data-word-id="{obj.id}" class="audio-check-status-select" style="margin-top: 8px;">
-                {options}
-            </select>
-            """
-
-        return mark_safe(html)
-
-    list_audio.short_description = _("audio")  # type: ignore[attr-defined]
-
-    def list_image(self, obj: Word) -> SafeString:
-        """
-        Generate HTML for displaying the word's images with controls in the admin list view.
-
-        This method creates HTML that includes the word's main image and unit-specific images,
-        along with controls for adding, replacing, or deleting images, and dropdowns for
-        the image check status.
-
-        Args:
-            obj: The word object
-
-        Returns:
-            str: HTML markup for displaying the word's images with controls
-        """
-        # Generate the word's main image container
-        word_image_container = self._generate_word_image_container(obj)
-
-        # Generate the unit-specific image containers
-        unit_word_images = self._generate_unit_word_images(obj)
-
-        # Combine all images into a single container
-        all_images = f'<div class="all-images-container"><div>{word_image_container}</div><div>{unit_word_images}</div></div>'
-
-        return mark_safe(all_images)
-
-    def list_example_sentence(self, obj: Word) -> SafeString:
-        """
-        Generate HTML for displaying the word's example sentence with controls in the admin list view.
-
-        This method creates HTML that includes the word's example sentence (as text),
-        along with controls for listening to the audio, uploading new audio and dropdowns for the check status.
-
-        Args:
-            obj: The word object
-
-        Returns:
-            str: HTML markup for displaying the word's example sentence with controls
-        """
-        example_sentence_html = self._generate_word_example_sentence_container(obj)
-        example_sentence_audio_html = (
-            self._generate_word_example_sentence_audio_container(obj)
-        )
-        example_sentence_combined = f'<div class="example-sentence-container"><div>{example_sentence_html}</div><div>{example_sentence_audio_html}</div></div>'
-        return mark_safe(example_sentence_combined)
-
-    list_example_sentence.short_description = _("Example sentence")  # type: ignore[attr-defined]
-
-    def _generate_example_sentence_text_html(
-        self, sentence: str, truncate_at: int = 40
-    ) -> str:
-        """Generate HTML for an example sentence, truncated with an expandable
-        "more" toggle if it exceeds `truncate_at` characters.
-
-        Args:
-            sentence: The raw example sentence text
-            truncate_at: The character count above which the sentence gets truncated
-
-        Returns:
-            str: HTML markup for the (optionally truncated) example sentence
-        """
-        if len(sentence) <= truncate_at:
-            return f"""<div class="example-sentence-hover-container">
-                <span>{escape(sentence)}</span>
-            </div>"""
-
-        truncated = escape(sentence[:truncate_at])
-        full = escape(sentence)
-        return f"""<div class="example-sentence-hover-container py-2">
-            <details class="example-sentence-details">
-                <summary class="example-sentence-summary">
-                    <span class="example-sentence-truncated">{truncated}&hellip;</span>
-                    <span class="example-sentence-full">{full}</span>
-                    <span class="example-sentence-more">{_("more")}</span>
-                </summary>
-            </details>
-        </div>"""
-
-    def _generate_word_example_sentence_container(self, obj: Word) -> str:
-        """Generate HTML for the word's example sentence container.
-
-        Args:
-            obj: The word object
-
-        Returns:
-            str: HTML markup for the word's example sentence container"""
-
-        example_sentence_html = (
-            self._generate_example_sentence_text_html(obj.example_sentence)
-            if obj.example_sentence
-            else ""
-        )
-
-        controls_html = f"""
-        <div class="example-sentence-controls" data-word-id="{obj.id}">
-            <button disabled type="button" class="edit-example-sentence-btn" style="display: {'inline-flex' if obj.example_sentence else 'none'};">
-                <span class="example-sentence-edit">✎</span>
-            </button>
-            <button disabled type="button" class="replace-example-sentence-btn" style="display: {'inline-flex' if obj.example_sentence else 'none'};">
-                <span class="example-sentence-replace">↻</span>
-            </button>
-            <button disabled type="button" class="delete-example-sentence-btn" style="display: {'inline-flex' if obj.example_sentence else 'none'};">
-                <span class="example-sentence-delete">×</span>
-            </button>
-            <input type="file" class="example-sentence-file-input" style="display: none;" accept="example_sentence/*">
-        </div>
-        """
-
-        word_options = ""
-        for value, display in CheckStatus.choices:
-            selected = "selected" if obj.example_sentence_check_status == value else ""
-            word_options += f'<option value="{value}" {selected}>{display}</option>'
-
-        word_example_sentence_check_status_html = f"""
-        <select name="example_sentence_check_status_{obj.id}" data-word-id="{obj.id}" class="example-sentence-check-status-select" style="margin-top: 8px;">
-            {word_options}
-        </select>
-        """
-
-        html = f'<div class="word-example-sentence-container">{example_sentence_html}{controls_html}</div>'
-        if obj.example_sentence:
-            html += word_example_sentence_check_status_html
-
-        return html
-
-    def _generate_word_example_sentence_audio_container(self, obj: Word) -> str:
-        """Generate HTML for the word's example sentence audio container.
-
-        Args:
-            obj: The word object
-
-        Returns:
-            str: HTML markup for the word's example sentence audio container"""
-
-        audio_html = ""
-        if obj.example_sentence_audio:
-            audio_html = f"""
-            <div class="audio-player-container">
-                <audio class="minimal-audio-player"><source src="{cache_busted_url(obj.example_sentence_audio)}" type="audio/mpeg"></audio>
-                <div class="play-btn">
-                    <div>
-                        <i class="fas fa-play"></i>
-                    </div>
-                </div>
-                <div class="pause-btn" style="display: none;">
-                    <div>
-                        <i class="fas fa-pause"></i>
-                    </div>
-                </div>
-            </div>
-            """
-
-        controls_html = f"""
-        <div class="example-sentence-audio-controls" data-word-id="{obj.id}">
-            <button disabled type="button" class="add-example-sentence-audio-btn" style="display: {'none' if obj.example_sentence_audio else 'inline-flex'};">
-                <span class="example-sentence-audio-add">+</span>
-            </button>
-            <button disabled type="button" class="replace-example-sentence-audio-btn" style="display: {'inline-flex' if obj.example_sentence_audio else 'none'};">
-                <span class="example-sentence-audio-replace">↻</span>
-            </button>
-            <button disabled type="button" class="delete-example-sentence-audio-btn" style="display: {'inline-flex' if obj.example_sentence_audio else 'none'};">
-                <span class="example-sentence-audio-delete">×</span>
-            </button>
-            <input type="file" class="example-sentence-audio-file-input" style="display: none;" accept="audio/*">
-        </div>
-        """
-
-        return f'<div class="word-example-sentence-audio-container py-2">{audio_html}{controls_html}</div>'
-
-    def _generate_word_image_container(self, obj: Word) -> str:
-        """
-        Generate HTML for the word's main image container.
-
-        Args:
-            obj: The word object
-
-        Returns:
-            str: HTML markup for the word's main image container
-        """
-        if obj.image:
-            image_html = f"""<div class="image-hover-container">
-                <a href="{escape(f"{settings.MEDIA_URL}{obj.image}")}" target="_blank">{get_image_tag(obj.image, width=50)}</a>
-                <div class="image-hover-overlay">
-                    <img src="{escape(f"{settings.MEDIA_URL}{obj.image}")}" alt="{escape(obj.word)}">
-                </div>
-            </div>"""
-        else:
-            image_html = ""
-
-        controls_html = f"""
-        <div class="image-controls" data-word-id="{obj.id}">
-            <button type="button" class="add-image-btn" style="display: {'none' if obj.image else 'inline-flex'};">
-                <span class="image-add">+</span>
-            </button>
-            <button type="button" class="replace-image-btn" style="display: {'inline-flex' if obj.image else 'none'};">
-                <span class="image-replace">↻</span>
-            </button>
-            <button type="button" class="delete-image-btn" style="display: {'inline-flex' if obj.image else 'none'};">
-                <span class="image-delete">×</span>
-            </button>
-            <input type="file" class="image-file-input" style="display: none;" accept="image/*">
-        </div>
-        """
-
-        word_options = ""
-        for value, display in CheckStatus.choices:
-            selected = "selected" if obj.image_check_status == value else ""
-            word_options += f'<option value="{value}" {selected}>{display}</option>'
-
-        word_image_check_status_html = f"""
-        <select name="image_check_status_{obj.id}" data-word-id="{obj.id}" class="image-check-status-select" style="margin-top: 8px;">
-            {word_options}
-        </select>
-        """
-
-        html = f'<div class="word-image-container">{image_html}{controls_html}</div>'
-        if obj.image:
-            html += word_image_check_status_html
-
-        return html
-
-    def _generate_unit_word_images(self, obj: Word) -> str:
-        """
-        Generate HTML for the unit-specific image containers.
-
-        Args:
-            obj: The word object
-
-        Returns:
-            str: HTML markup for all unit-specific image containers
-        """
-        unit_word_images = ""
-
-        for relation in obj.unit_word_relations.select_related("unit").all():
-            unit_word_images += self._generate_unit_word_image(relation)
-
-        return unit_word_images
-
-    def _generate_unit_word_image(self, relation: UnitWordRelation) -> str:
-        """
-        Generate HTML for a single unit-word image container.
-
-        Args:
-            relation: The UnitWordRelation object
-
-        Returns:
-            str: HTML markup for a single unit-word image container
-        """
-        unit_name = relation.unit.title
-        if relation.image:
-            unit_image_html = f"""<div class="image-hover-container">
-                <a href="{escape(f"{settings.MEDIA_URL}{relation.image}")}" target="_blank">{get_image_tag(relation.image, width=50)}</a>
-                <div class="image-hover-overlay">
-                    <img src="{escape(f"{settings.MEDIA_URL}{relation.image}")}" alt="{escape(relation.unit.title)}">
-                </div>
-            </div>"""
-        else:
-            unit_image_html = ""
-
-        unit_controls_html = f"""
-        <div class="unitword-image-controls" data-unitword-id="{relation.id}">
-            <button type="button" class="add-unitword-image-btn" style="display: {'none' if relation.image else 'inline-flex'};">
-                <span class="unitword-image-add">+</span>
-            </button>
-            <button type="button" class="replace-unitword-image-btn" style="display: {'inline-flex' if relation.image else 'none'};">
-                <span class="unitword-image-replace">↻</span>
-            </button>
-            <button type="button" class="delete-unitword-image-btn" style="display: {'inline-flex' if relation.image else 'none'};">
-                <span class="unitword-image-delete">×</span>
-            </button>
-            <input type="file" class="unitword-image-file-input" style="display: none;" accept="image/*">
-        </div>
-        """
-
-        unit_options = ""
-        for value, display in CheckStatus.choices:
-            selected = "selected" if relation.image_check_status == value else ""
-            unit_options += f'<option value="{value}" {selected}>{display}</option>'
-
-        unit_image_check_status_html = f"""
-        <select name="unitword_image_check_status_{relation.id}" data-unitword-id="{relation.id}" class="unitword-image-check-status-select" style="margin-top: 8px;">
-            {unit_options}
-        </select>
-        """
-
-        unit_name_html = f'<div class="unit-name">{unit_name}</div>'
-
-        html = f"""
-        <div class="unitword-image-wrapper">
-            {unit_name_html}
-            <div class="unitword-image-container">
-                {unit_image_html}
-                {unit_controls_html}
-            </div>
-        </div>
-        """
-
-        if relation.image:
-            html += unit_image_check_status_html
-
-        return html
-
-    list_image.short_description = _("Image")  # type: ignore[attr-defined]
 
     def singular_article_display(self, obj: Word) -> str:
         """
